chore(scan): separate scan function into separate file
